+++ lib/fast_json_selector.dart
@@ -1,15 +1,13 @@
 import 'fast_json_handler.dart' as parser;
-import 'fast_json_handler.dart' show JsonEvent, JsonParserHandler;
+import 'fast_json_handler.dart' show JsonHandlerEvent, JsonParserHandler;
 
-class FastJsonSelector {
-  void parse(String source,
-      {required void Function(FastJsonSelectorEvent context) select}) {
-    final handler = _JsonParserHandler(select);
-    parser.parse(source, handler);
-  }
+void parse(String source,
+    {required void Function(JsonSelectorEvent context) select}) {
+  final handler = _JsonParserHandler(select);
+  parser.parse(source, handler);
 }
 
-class FastJsonSelectorEvent {
+class JsonSelectorEvent {
   final List buffer = [];
 
   dynamic lastValue;
@@ -18,50 +16,50 @@ class FastJsonSelectorEvent {
 }
 
 class _JsonParserHandler<T> extends JsonParserHandler {
-  final void Function(FastJsonSelectorEvent context) _select;
+  final void Function(JsonSelectorEvent context) _select;
 
-  final FastJsonSelectorEvent context = FastJsonSelectorEvent();
+  final JsonSelectorEvent context = JsonSelectorEvent();
 
   _JsonParserHandler(this._select);
 
   @override
-  void handle(JsonEvent event, dynamic value) {
+  void handle(JsonHandlerEvent event, dynamic value) {
     switch (event) {
-      case JsonEvent.beginArray:
+      case JsonHandlerEvent.beginArray:
         context.buffer.add([]);
         context.levels.add('[]');
         context.levels.add(0);
         break;
-      case JsonEvent.beginObject:
+      case JsonHandlerEvent.beginObject:
         context.buffer.add(<String, dynamic>{});
         context.levels.add('{}');
         break;
-      case JsonEvent.endArray:
+      case JsonHandlerEvent.endArray:
         context.levels.removeLast();
         context.lastValue = context.buffer.removeLast();
         _select(context);
         context.levels.removeLast();
         break;
-      case JsonEvent.endObject:
+      case JsonHandlerEvent.endObject:
         context.lastValue = context.buffer.removeLast();
         _select(context);
         context.levels.removeLast();
         break;
-      case JsonEvent.element:
+      case JsonHandlerEvent.element:
         _select(context);
         context.buffer.last.add(context.lastValue);
         context.levels.last++;
         break;
-      case JsonEvent.beginKey:
+      case JsonHandlerEvent.beginKey:
         context.levels.add(context.lastValue as String);
         break;
-      case JsonEvent.endKey:
+      case JsonHandlerEvent.endKey:
         _select(context);
         context.buffer.last[value] = context.lastValue;
         context.levels.removeLast();
 
         break;
-      case JsonEvent.value:
+      case JsonHandlerEvent.value:
         context.lastValue = value;
         break;
     }